Core: fixed a few warnings.

+++ core/src/Spirit/IO.cpp
@@ -96,7 +96,7 @@ void IO_Chain_Read(State * state, const char * file, int idx_image, int idx_chai
 	chain->Unlock();
 
 	// Update llg simulation information array size
-	if (state->simulation_information_llg[idx_chain].size() < chain->noi)
+	if ((int)state->simulation_information_llg[idx_chain].size() < chain->noi)
 	{
 		for (int i=state->simulation_information_llg[idx_chain].size(); i < chain->noi; ++i)
 			state->simulation_information_llg[idx_chain].push_back(std::shared_ptr<Simulation_Information>(new Simulation_Information()));
